% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nor.min.max.R
\name{nor.min.max}
\alias{nor.min.max}
\title{Function to normalize data to bring values into alignment. This function
uses feature scaling to normalize values in a dataset between 0 and 1.}
\usage{
nor.min.max(x)
}
\arguments{
\item{x}{An integer object of numeric value}
}
\value{
Returns a numeric value normalized between 0 and 1.
}
\description{
Function to normalize data to bring values into alignment. This function
uses feature scaling to normalize values in a dataset between 0 and 1.
}
\examples{

# Load sample dataset
data(iris)
# View sample matrix
iris[1:5, 1:5]
# Coerce sample matrix to numeric values
iris <- t(apply(iris[, 1:4], 1, as.numeric))
#Normalize values in the matrix using the function
data.min.max <- t(apply(iris, 1, nor.min.max))
}
\author{
Peiyong Guan
}

